Content end point logic change
A modification was made to the content enpoint of the API to eliminate
the 301 redirect method. It is presumed that it will be more efficient
as well as solve a discovered issue preventing the Articulate Mobile
Player from being able to access an embedded video when requested from
the content end point.